requirements.txt
attrs==20.3.0
blessed==1.17.6
botocore==1.20.79
certifi==2020.12.5
cffi==1.14.5
chalice==1.23.0
chardet==4.0.0
click==7.1.2
future==0.18.2
idna==2.10
inquirer==2.7.0
jmespath==0.10.0
line-bot-sdk==1.19.0
mypy-extensions==0.4.3
opencc-python-reimplemented==0.1.6
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.4.1
readchar==2.0.1
requests==2.25.1
six==1.16.0
urllib3==1.26.4
wcwidth==0.2.5
